BASW-646: Fix Primary location type mapping on civi profiles
If the user is configured an event with a profile that contain a
contact Primary address then whatever you  insert into the postal
code won't fill the other address fields.

This happen because the Primary location type is a special hardcoded
type in CiviCRM and not a location type that stored in the location
types table, where the JS code here tries to treat it as a normal
location type with normal Id and thus failing to find it in the DOM
tree.